backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
pydantic>=2.6
pydantic[email]>=2.6
python-dotenv>=1.0
httpx>=0.27

# Database
SQLAlchemy>=2.0
asyncpg>=0.29

# Optional (recommended) for migrations
alembic>=1.13
psycopg[binary]>=3.1
